requirements.txt
# Emmanuel Ato Gaisie
# 10022300173
# CS4241

numpy
pandas
PyPDF2
sentence-transformers
faiss-cpu
streamlit
